+++ front-end/package.json
@@ -6,6 +6,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.30",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
+    "@headlessui/react": "^0.2.0",
     "axios": "^0.18.0",
     "camelcase-keys": "^6.2.2",
     "connected-react-router": "^6.1.0",
